requirements.txt
customtkinter
kneed
matplotlib
numpy
opencv_python
Pillow
scikit_learn
scipy
scikit-image
